Create UserGroup if not exists

+++ UJCSystem/Vokabular.MainService.Core/Works/Permission/DeleteRoleWork.cs
@@ -21,9 +21,12 @@ namespace Vokabular.MainService.Core.Works.Permission
         protected override void ExecuteWorkImplementation()
         {
             var group = m_permissionRepository.FindGroupByExternalId(m_roleId);
-            m_permissionRepository.Delete(group);
-            m_permissionRepository.Flush();
-
+            if (group != null)
+            {
+                m_permissionRepository.Delete(group);
+                m_permissionRepository.Flush();
+            }
+            
             var client = m_communicationProvider.GetAuthRoleApiClient();
             client.HttpClient.DeleteItemAsync<RoleContract>(m_roleId).GetAwaiter().GetResult();
         }
